+++ main.rb
@@ -1,4 +1,4 @@
-require_relative 'app'
+require_relative 'start'
 
 class Main
   def initialize
@@ -18,24 +18,6 @@ class Main
       print "6 - List all rentals for a given person id\n"
       print "7 - Exit\n"
       start_app
-    end
-  end
-
-  def start_app
-    @output = gets.chomp.to_i
-    case @output
-    when 1
-      @app.list_books
-    when 2
-      @app.list_people
-    when 3
-      @app.create_person
-    when 4
-      @app.create_book
-    when 5
-      @app.create_rental
-    when 6
-      @app.list_rentals
     end
   end
 end
